Add files via upload
This is a snake-water-gun game where one player is the user and the other player is cpu of the computer. The user has to either enter capital S (snake) or W (water) or G (gun) and the cpu will randomly generate  its choice. Incase of Snake and Water, Snake wins. Between Snake and Gun, Gun wins and between Water and Gun, Water wins.